requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
websockets==12.0
twilio==8.12.0
supabase==2.3.0
python-dateutil==2.8.2 